Make ui tests std-only, write separate no-std tests
