+++ .github/workflows/ci.yml
@@ -72,6 +72,11 @@ jobs:
           toolchain: ${{ matrix.rust }}
           override: true
       - uses: Swatinem/rust-cache@v1
+
+      - name: Downgrade deps for MSRV
+        run: |
+          cargo update -p libc --precise 0.2.160
+          cargo update -p pretty_assertions --precise 1.4.0
       - uses: actions-rs/cargo@v1
         with:
           command: test
