--- tests/no_std/without.stderr
@@ -1,22 +1,0 @@
-warning: unused import: `displaydoc::Display`
-  --> $DIR/without.rs:20:5
-   |
-20 | use displaydoc::Display;
-   |     ^^^^^^^^^^^^^^^^^^^
-   |
-   = note: `#[warn(unused_imports)]` on by default
-
-error[E0277]: `FakeType` doesn't implement `Display`
-  --> $DIR/without.rs:25:37
-   |
-25 | static_assertions::assert_impl_all!(FakeType: core::fmt::Display);
-   |                                     ^^^^^^^^ `FakeType` cannot be formatted with the default formatter
-   |
-   = help: the trait `Display` is not implemented for `FakeType`
-   = note: in format strings you may be able to use `{:?}` (or {:#?} for pretty-print) instead
-note: required by a bound in `assert_impl_all`
-  --> $DIR/without.rs:25:1
-   |
-25 | static_assertions::assert_impl_all!(FakeType: core::fmt::Display);
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `assert_impl_all`
-   = note: this error originates in the macro `static_assertions::assert_impl_all` (in Nightly builds, run with -Z macro-backtrace for more info)
