--- tests/std/multi_line.rs
@@ -1,1 +1,0 @@
-../no_std/multi_line.rs
