+++ tests/ui/without.stderr
@@ -1,0 +1,22 @@
+warning: unused import: `displaydoc::Display`
+ --> tests/ui/without.rs:1:5
+  |
+1 | use displaydoc::Display;
+  |     ^^^^^^^^^^^^^^^^^^^
+  |
+  = note: `#[warn(unused_imports)]` on by default
+
+error[E0277]: `FakeType` doesn't implement `std::fmt::Display`
+ --> tests/ui/without.rs:6:37
+  |
+6 | static_assertions::assert_impl_all!(FakeType: core::fmt::Display);
+  |                                     ^^^^^^^^ `FakeType` cannot be formatted with the default formatter
+  |
+  = help: the trait `std::fmt::Display` is not implemented for `FakeType`
+  = note: in format strings you may be able to use `{:?}` (or {:#?} for pretty-print) instead
+note: required by a bound in `assert_impl_all`
+ --> tests/ui/without.rs:6:1
+  |
+6 | static_assertions::assert_impl_all!(FakeType: core::fmt::Display);
+  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ required by this bound in `assert_impl_all`
+  = note: this error originates in the macro `static_assertions::assert_impl_all` (in Nightly builds, run with -Z macro-backtrace for more info)
